+++ src/extension/itunes/itunes_channel_extension.rs
@@ -27,7 +27,7 @@ use crate::toxml::{ToXml, WriterExt};
     builder(
         setter(into),
         default,
-        build_fn(name = "build_impl", private, error = "never::Never")
+        build_fn(name = "build_impl", private, error = "std::convert::Infallible")
     )
 )]
 pub struct ITunesChannelExtension {