+++ src/extension/itunes/itunes_item_extension.rs
@@ -26,7 +26,7 @@ use crate::toxml::{ToXml, WriterExt};
     builder(
         setter(into),
         default,
-        build_fn(name = "build_impl", private, error = "never::Never")
+        build_fn(name = "build_impl", private, error = "std::convert::Infallible")
     )
 )]
 pub struct ITunesItemExtension {