+++ src/extension/syndication.rs
@@ -71,7 +71,7 @@ impl fmt::Display for UpdatePeriod {
     builder(
         setter(into),
         default,
-        build_fn(name = "build_impl", private, error = "never::Never")
+        build_fn(name = "build_impl", private, error = "std::convert::Infallible")
     )
 )]
 pub struct SyndicationExtension {